+++ R/print-bvharspec.R
@@ -14,7 +14,7 @@ print.bvharspec <- function(x, digits = max(3L, getOption("digits") - 3L), ...) 
     "Flat" = "?bvar_flat",
     "MN_VAR" = "?bvhar_minnesota",
     "MN_VHAR" = "?bvhar_minnesota",
-    "Hierarchical" = "?bvar_",
+    "MN_Hierarchical" = "?bvar_",
     stop("Invalid 'x$prior' element")
   )
   cat(paste0("# Type '", fit_func, "' in the console for some help.", "\n"))