+++ NEWS.md
@@ -1,3 +1,15 @@
+# bvhar 0.12.0
+
+* Now can use heavy-tailed distribution ([Multivariate t-distribution](https://en.wikipedia.org/wiki/Multivariate_t-distribution)) when generating VAR and VHAR process (`sim_var()` and `sim_vhar()`).
+
+* Also provide independent MVT generation function (`sim_mvt()`).
+
+# bvhar 0.11.0
+
+* Added `method = c("nor", "chol", "qr")` option in VAR and VHAR fitting function to use cholesky and Householder QR method (`var_lm()` and `vhar_lm()`).
+
+* Now `include_mean` works internally with `Rcpp`.
+
 # bvhar 0.10.0
 
 * Add partial t-test for each VAR and VHAR coefficient (`summary.varlse()` and `summary.vharlse()`).